valuation-mvp/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-multipart==0.0.20
openai==1.59.9
Pillow==11.1.0
lxml==5.3.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
blocket-api==0.4.32
slowapi==0.1.9
